Copilot/replace GLFW, Nanovg and Miniaudio with SDL3 (#9)
* docs: add SDL3 migration plan for replacing nanovg and miniaudio

Agent-Logs-Url: https://github.com/acoto87/war1/sessions/e66b52bb-da8d-4ab2-9f08-d62b8381544d


* Refactor rendering system to use SDL and improve color handling

- Replaced NVG rendering functions with SDL rendering functions.
- Introduced a new color system using u8Color instead of NVGcolor.
- Added render state management for transformations and alpha blending.
- Updated sprite handling to use SDL_Texture instead of NVG images.
- Refactored rendering functions to accommodate new rendering architecture.
- Changed event handling for quitting the application to use SDL events.
- Removed NVGcontext references and related functions from the codebase.
- Improved performance by reducing unnecessary state saves/restores.
- Updated math header to include pragma once for better header management.
- Cleaned up unused variables and code for better readability.

* Remove previous build scripts.

* Refactor and enhance various components for improved stability and performance

- Updated `war_scene_download.c` to ensure proper type casting for UI text bounding dimensions.
- Enhanced error logging in `updateSceneDownload` to provide more informative messages.
- Added unused parameter macros in multiple state machine files to suppress compiler warnings.
- Changed sprite frame dimensions from `u8` to `u16` in `war_types.h` for better compatibility.
- Ensured proper type casting in sprite creation functions to prevent potential overflows.
- Improved pathfinding and entity management in state machine files for better performance.
- Updated UI handling in `war_ui.c` to ensure correct type casting for width settings.
- Renamed upgrade and command data arrays for clarity in `war_units.h`.
- Added assertions to validate dimensions in sprite creation to prevent invalid states.
- General code cleanup and consistency improvements across multiple files.

* Fix compiler warnings.

* Remove sdl3-migration-plan.md file.

---------

+++ src/war_scene_briefing.c
@@ -37,7 +37,7 @@ void enterSceneBriefingHumans(WarContext* context)
     WarEntity* briefingText = createUIText(context, "txtBriefing", 1, 10, data.briefingText, vec2i(20, 160));
     setUITextColor(briefingText, u8RgbColor(255, 215, 138));
     setUITextMultiline(briefingText, true);
-    setUITextBoundings(briefingText, vec2f(context->originalWindowWidth - 40, 200));
+    setUITextBoundings(briefingText, vec2f((f32)(context->originalWindowWidth - 40), 200.0f));
     setUITextHorizontalAlign(briefingText, WAR_TEXT_ALIGN_LEFT);
     setUITextVerticalAlign(briefingText, WAR_TEXT_ALIGN_TOP);
     setUITextLineAlign(briefingText, WAR_TEXT_ALIGN_LEFT);
@@ -86,7 +86,7 @@ void enterSceneBriefingOrcs(WarContext* context)
     WarEntity* briefingText = createUIText(context, "txtBriefing", 1, 10, data.briefingText, vec2i(20, 160));
     setUITextColor(briefingText, u8RgbColor(255, 215, 138));
     setUITextMultiline(briefingText, true);
-    setUITextBoundings(briefingText, vec2f(context->originalWindowWidth - 40, 200));
+    setUITextBoundings(briefingText, vec2f((f32)(context->originalWindowWidth - 40), 200.0f));
     setUITextHorizontalAlign(briefingText, WAR_TEXT_ALIGN_LEFT);
     setUITextVerticalAlign(briefingText, WAR_TEXT_ALIGN_TOP);
     setUITextLineAlign(briefingText, WAR_TEXT_ALIGN_LEFT);
